Add tests for duplicate alias detection and multiple imports
- Implemented a test case to verify that duplicate aliases for the same module are correctly detected and reported as errors.
- Created a test to demonstrate that importing the same module multiple times with different aliases works correctly, allowing access to the same constants and functions through all aliases.
- Updated AST representations for the new test cases to ensure accurate parsing and validation of imports and aliases.
