+++ internal/semantics/collector/importAliasing.go
@@ -3,6 +3,7 @@ package collector
 import (
 	"compiler/internal/context_v2"
 	"compiler/internal/diagnostics"
+	"compiler/internal/source"
 	"fmt"
 	"strings"
 )
@@ -11,43 +12,56 @@ import (
 // This enables module::symbol resolution in the resolver phase
 func buildImportAliasMap(ctx *context_v2.CompilerContext, mod *context_v2.Module) {
 	// Track where each alias was first defined for better error reporting
-	aliasLocations := make(map[string]*context_v2.Import)
+	// Map: aliasName -> {importPath, location}
+	aliasInfo := make(map[string]struct {
+		importPath string
+		location   *source.Location
+	})
 
 	for _, imp := range mod.Imports {
-		var aliasName string
+		// Process each alias for this import
+		for i, alias := range imp.Aliases {
+			var aliasName string
 
-		// If there's an alias, use it
-		if imp.Alias != "" {
-			aliasName = imp.Alias
-		} else {
-			// Extract the last component of the import path as the default name
-			// e.g., "test_project/utils" -> "utils"
-			//       "std/math" -> "math"
-			parts := splitImportPath(imp.Path)
-			if len(parts) > 0 {
-				aliasName = parts[len(parts)-1]
+			// If there's an explicit alias, use it
+			if alias != "" {
+				aliasName = alias
 			} else {
-				// Skip invalid import paths
+				// Extract the last component of the import path as the default name
+				// e.g., "test_project/utils" -> "utils"
+				//       "std/math" -> "math"
+				parts := splitImportPath(imp.Path)
+				if len(parts) > 0 {
+					aliasName = parts[len(parts)-1]
+				} else {
+					// Skip invalid import paths
+					continue
+				}
+			}
+
+			// Check for duplicate alias
+			if existing, exists := aliasInfo[aliasName]; exists {
+				// Report error for duplicate alias
+				ctx.Diagnostics.Add(
+					diagnostics.NewError(fmt.Sprintf("duplicate import alias '%s'", aliasName)).
+						WithCode("C-DUP-ALIAS").
+						WithPrimaryLabel(mod.FilePath, imp.Locations[i], fmt.Sprintf("'%s' already used for another import", aliasName)).
+						WithSecondaryLabel(mod.FilePath, existing.location, "previous import here").
+						WithHelp(fmt.Sprintf("use a different alias: import \"%s\" as %s_alt", imp.Path, aliasName)),
+				)
 				continue
 			}
-		}
 
-		// Check for duplicate alias
-		if existingImport, exists := aliasLocations[aliasName]; exists {
-			// Report error for duplicate alias
-			ctx.Diagnostics.Add(
-				diagnostics.NewError(fmt.Sprintf("duplicate import alias '%s'", aliasName)).
-					WithCode("C-DUP-ALIAS").
-					WithPrimaryLabel(mod.FilePath, imp.Location, fmt.Sprintf("'%s' already used for another import", aliasName)).
-					WithSecondaryLabel(mod.FilePath, existingImport.Location, "previous import here").
-					WithHelp(fmt.Sprintf("use a different alias: import \"%s\" as %s_alt", imp.Path, aliasName)),
-			)
-			continue
+			// Add to map
+			mod.ImportAliasMap[aliasName] = imp.Path
+			aliasInfo[aliasName] = struct {
+				importPath string
+				location   *source.Location
+			}{
+				importPath: imp.Path,
+				location:   imp.Locations[i],
+			}
 		}
-
-		// Add to map
-		mod.ImportAliasMap[aliasName] = imp.Path
-		aliasLocations[aliasName] = imp
 	}
 }
