+++ test_project/test_duplicate_alias_error.fer.ast.json
@@ -1,0 +1,212 @@
+{
+  "FullPath": "/home/fuad/Documents/dev/go/Ferret/test_project/test_duplicate_alias_error.fer",
+  "ImportPath": "",
+  "Alias": "",
+  "Nodes": [
+    {
+      "Path": {
+        "Kind": 2,
+        "Value": "test_project/utils",
+        "Start": {
+          "Line": 4,
+          "Column": 8,
+          "Index": 105
+        },
+        "End": {
+          "Line": 4,
+          "Column": 28,
+          "Index": 125
+        }
+      },
+      "Alias": {
+        "Name": "U",
+        "Start": {
+          "Line": 4,
+          "Column": 32,
+          "Index": 129
+        },
+        "End": {
+          "Line": 4,
+          "Column": 33,
+          "Index": 130
+        }
+      },
+      "LocationOnDisk": "",
+      "Start": {
+        "Line": 4,
+        "Column": 1,
+        "Index": 98
+      },
+      "End": {
+        "Line": 4,
+        "Column": 34,
+        "Index": 131
+      }
+    },
+    {
+      "Path": {
+        "Kind": 2,
+        "Value": "test_project/utils",
+        "Start": {
+          "Line": 5,
+          "Column": 8,
+          "Index": 139
+        },
+        "End": {
+          "Line": 5,
+          "Column": 28,
+          "Index": 159
+        }
+      },
+      "Alias": {
+        "Name": "U",
+        "Start": {
+          "Line": 5,
+          "Column": 32,
+          "Index": 163
+        },
+        "End": {
+          "Line": 5,
+          "Column": 33,
+          "Index": 164
+        }
+      },
+      "LocationOnDisk": "",
+      "Start": {
+        "Line": 5,
+        "Column": 1,
+        "Index": 132
+      },
+      "End": {
+        "Line": 5,
+        "Column": 34,
+        "Index": 165
+      }
+    },
+    {
+      "Name": {
+        "Name": "main",
+        "Start": {
+          "Line": 7,
+          "Column": 4,
+          "Index": 201
+        },
+        "End": {
+          "Line": 7,
+          "Column": 8,
+          "Index": 205
+        }
+      },
+      "Type": {
+        "Params": null,
+        "Result": null,
+        "Start": {
+          "Line": 7,
+          "Column": 1,
+          "Index": 198
+        },
+        "End": {
+          "Line": 7,
+          "Column": 10,
+          "Index": 207
+        }
+      },
+      "Body": {
+        "Nodes": [
+          {
+            "Decls": [
+              {
+                "Name": {
+                  "Name": "x",
+                  "Start": {
+                    "Line": 8,
+                    "Column": 9,
+                    "Index": 218
+                  },
+                  "End": {
+                    "Line": 8,
+                    "Column": 10,
+                    "Index": 219
+                  }
+                },
+                "Type": null,
+                "Value": {
+                  "X": {
+                    "Name": "U",
+                    "Start": {
+                      "Line": 8,
+                      "Column": 14,
+                      "Index": 223
+                    },
+                    "End": {
+                      "Line": 8,
+                      "Column": 15,
+                      "Index": 224
+                    }
+                  },
+                  "Selector": {
+                    "Name": "MAX",
+                    "Start": {
+                      "Line": 8,
+                      "Column": 17,
+                      "Index": 226
+                    },
+                    "End": {
+                      "Line": 8,
+                      "Column": 20,
+                      "Index": 229
+                    }
+                  },
+                  "Start": {
+                    "Line": 8,
+                    "Column": 14,
+                    "Index": 223
+                  },
+                  "End": {
+                    "Line": 8,
+                    "Column": 20,
+                    "Index": 229
+                  }
+                }
+              }
+            ],
+            "Start": {
+              "Line": 8,
+              "Column": 5,
+              "Index": 214
+            },
+            "End": {
+              "Line": 8,
+              "Column": 21,
+              "Index": 230
+            }
+          }
+        ],
+        "Scope": null,
+        "Start": {
+          "Line": 7,
+          "Column": 11,
+          "Index": 208
+        },
+        "End": {
+          "Line": 9,
+          "Column": 2,
+          "Index": 232
+        }
+      },
+      "Scope": null,
+      "Start": {
+        "Line": 7,
+        "Column": 1,
+        "Index": 198
+      },
+      "End": {
+        "Line": 9,
+        "Column": 2,
+        "Index": 232
+      }
+    }
+  ],
+  "Start": null,
+  "End": null
+}
